requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
sqlparse==0.4.4
tzdata==2024.1